Move server.js and start-server.js to src/frame (#45767)
