+++ .github/workflows/check-broken-links-github-github.yml
@@ -60,7 +60,7 @@ jobs:
           ENABLED_LANGUAGES: en
         run: |
 
-          node server.js &
+          node src/frame/server.js &
           sleep 5
           curl --retry-connrefused --retry 3 -I http://localhost:4000/
