+++ package.json
@@ -14,10 +14,10 @@
       "default": "./src/*"
     }
   },
-  "exports": "./server.js",
+  "exports": "./src/frame/server.js",
   "scripts": {
     "build": "next build",
-    "debug": "cross-env NODE_ENV=development ENABLED_LANGUAGES=en nodemon --inspect server.js",
+    "debug": "cross-env NODE_ENV=development ENABLED_LANGUAGES=en nodemon --inspect src/frame/server.js",
     "dev": "cross-env npm start",
     "fixture-dev": "cross-env ROOT=tests/fixtures npm start",
     "fixture-test": "cross-env ROOT=tests/fixtures npm test -- tests/rendering-fixtures",
@@ -35,14 +35,14 @@
     "rest-dev": "node src/rest/scripts/update-files.js",
     "show-action-deps": "echo 'Action Dependencies:' && rg '^[\\s|-]*(uses:.*)$' .github -I -N --no-heading -r '$1$2' | sort | uniq | cut -c 7-",
     "prestart": "node src/workflows/cmp-files.js package-lock.json .installed.package-lock.json || npm install && cp package-lock.json .installed.package-lock.json",
-    "start": "cross-env NODE_ENV=development ENABLED_LANGUAGES=en nodemon server.js",
-    "start-all-languages": "cross-env NODE_ENV=development nodemon server.js",
-    "start-for-playwright": "cross-env ROOT=tests/fixtures TRANSLATIONS_FIXTURE_ROOT=tests/fixtures/translations ENABLED_LANGUAGES=en,ja NODE_ENV=test node server.js",
+    "start": "cross-env NODE_ENV=development ENABLED_LANGUAGES=en nodemon src/frame/server.js",
+    "start-all-languages": "cross-env NODE_ENV=development nodemon src/frame/server.js",
+    "start-for-playwright": "cross-env ROOT=tests/fixtures TRANSLATIONS_FIXTURE_ROOT=tests/fixtures/translations ENABLED_LANGUAGES=en,ja NODE_ENV=test node src/frame/server.js",
     "sync-rest": "node src/rest/scripts/update-files.js",
     "sync-search": "cross-env NODE_OPTIONS='--max_old_space_size=8192' start-server-and-test sync-search-server 4002 sync-search-indices",
     "sync-search-ghes-release": "cross-env GHES_RELEASE=1 start-server-and-test sync-search-server 4002 sync-search-indices",
     "sync-search-indices": "node src/search/scripts/sync-search-indices.js",
-    "sync-search-server": "cross-env NODE_ENV=production PORT=4002 MINIMAL_RENDER=true CHANGELOG_DISABLED=true node server.js",
+    "sync-search-server": "cross-env NODE_ENV=production PORT=4002 MINIMAL_RENDER=true CHANGELOG_DISABLED=true node src/frame/server.js",
     "sync-webhooks": "src/rest/scripts/update-files.js -o webhooks",
     "test": "cross-env NODE_OPTIONS='--max_old_space_size=4096 --experimental-vm-modules' jest --logHeapUsage",
     "test-watch": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --watch --notify --notifyMode=change --coverage",
